requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
idna==2.8
pymongo==3.8.0
python-dotenv==0.10.1
requests==2.21.0
soupsieve==1.9.1
urllib3==1.24.3
